address iss #37. add 3 new models and basic OLT logic including downstream resolution. also, cleaned up root models folder
