+++ netbox_netisp/netbox_netisp/views/generic.py
@@ -20,6 +20,8 @@ from django.contrib import messages
 import logging
 
 from netbox_netisp.models import *
+from netbox_netisp.netbox_netisp.models.wireless.models import *
+from netbox_netisp.netbox_netisp.models.crm.models import *
 
 
 class ObjectListView(View):
@@ -306,7 +308,9 @@ class HomeView(View):
         cpes = CustomerPremiseEquipment.objects.all()
         customers = Customer.objects.all()
         wirelesstickets = WirelessTicket.objects.all()
-
+        olts = OLT.objects.all()
+        splitters = GPONSplitter.objects.all()
+        onts = ONT.objects.all()
 
         return render(request, self.template_name, {
             "service_count": len(services),
@@ -317,5 +321,8 @@ class HomeView(View):
             "antenna_count": len(antennas),
             "cpe_count": len(cpes),
             "customer_count": len(customers),
-            "wirelessticket_count": len(wirelesstickets)
+            "wirelessticket_count": len(wirelesstickets),
+            "olt_count": len(olts),
+            "splitter_count": len(splitters),
+            "ont_count": len(onts),
         })
