+++ netbox_netisp/tables.py
@@ -6,6 +6,7 @@ from django_tables2.utils import A  # alias for Accessor
 
 from .models import *
 
+
 from utilities.tables import (
     BaseTable,
     ButtonsColumn,
@@ -13,6 +14,9 @@ from utilities.tables import (
     TagColumn,
     ToggleColumn
 )
+
+from .netbox_netisp.models.wireless.models import *
+from .netbox_netisp.models.crm.models import *
 
 
 
@@ -164,3 +168,24 @@ class AttachmentTable(BaseTable):
     class Meta(BaseTable.Meta):
         model = Attachment
         fields = ("pk", "image", "account", "service", "address")
+
+class OLTTable(BaseTable):
+    pk = tables.LinkColumn()
+
+    class Meta(BaseTable.Meta):
+        model = OLT
+        fields = ("pk", "name")
+
+class GPONSplitterTable(BaseTable):
+    pk = tables.LinkColumn()
+    uplink_type = tables.Column(verbose_name="Splitter Type")
+    class Meta(BaseTable.Meta):
+        model = GPONSplitter
+        fields = ("pk", "name", "uplink_type")
+
+class ONTTable(BaseTable):
+    pk = tables.LinkColumn()
+
+    class Meta(BaseTable.Meta):
+        model = ONT
+        fields = ("pk", "name", "uplink")
